+++ app/javascript/vue/tasks/collection_objects/slide_breakdown/components/Assign/CatalogNumber.vue
@@ -14,18 +14,18 @@
         </label>
       </li>
     </ul>
-    <div class="align-start margin-medium-top">
-      <smart-selector
+    <div
+      class="horizontal-left-content align-start margin-medium-top gap-small"
+    >
+      <SmartSelector
         model="namespaces"
         klass="CollectionObject"
         pin-section="Namespaces"
         pin-type="Namespace"
         @selected="setValue"
       />
-      <lock-component
-        class="margin-small-left"
-        v-model="lock.identifier"
-      />
+      <lock-component v-model="lock.identifier" />
+      <WidgetNamespace @create="setValue" />
     </div>
     <p
       v-if="identifier.namespace_id"
@@ -64,6 +64,7 @@
 
 <script>
 import SmartSelector from '@/components/ui/SmartSelector'
+import WidgetNamespace from '@/components/ui/Widget/WidgetNamespace.vue'
 import { GetterNames } from '../../store/getters/getters'
 import { MutationNames } from '../../store/mutations/mutations'
 import SharedComponent from '../shared/lock.js'
@@ -71,7 +72,8 @@ import SharedComponent from '../shared/lock.js'
 export default {
   mixins: [SharedComponent],
   components: {
-    SmartSelector
+    SmartSelector,
+    WidgetNamespace
   },
   computed: {
     identifier: {