+++ config/routes/data.rb
@@ -296,6 +296,7 @@ end
 # TODO: these should default json?
 resources :dwc_occurrences, only: [:create] do
   collection do
+    match :filter, to: 'dwc_occurrences#index', via: [:get, :post]
     get :index, defaults: {format: :json}
     get 'metadata', defaults: {format: :json}
     get 'predicates', defaults: {format: :json}