+++ config/routes/tasks.rb
@@ -5,6 +5,16 @@ scope :tasks do
     end
   end
 
+  scope :dwc_occurrences do
+    scope :filter, controller: 'tasks/dwc_occurrences/filter' do
+      get '/', action: :index, as: 'filter_dwc_occurrences_task'
+    end
+
+    scope :status, controller: 'tasks/dwc_occurrences/status' do
+      get '/', action: :index
+      post '/', action: :index
+    end
+  end
 
   scope :data_attributes do
     scope :field_synchronize, controller: 'tasks/data_attributes/field_synchronize' do